Move in_allowance rules into allowance.go
