+++ comp/networkpath/npcollector/impl/BUILD.bazel
@@ -6,6 +6,7 @@ load("//bazel/rules/go:dd_agent_go_test.bzl", "dd_agent_go_test")
 go_library(
     name = "impl",
     srcs = [
+        "allowance.go",
         "basic.go",
         "config.go",
         "localips.go",
