+++ comp/networkpath/npcollector/impl/npcollector.go
@@ -40,8 +40,7 @@ const (
 	reverseDNSLookupMetricPrefix        = common.NetworkPathCollectorMetricPrefix + "reverse_dns_lookup."
 	reverseDNSLookupFailuresMetricName  = reverseDNSLookupMetricPrefix + "failures"
 	reverseDNSLookupSuccessesMetricName = reverseDNSLookupMetricPrefix + "successes"
-	netpathConnsSkippedMetricName       = common.NetworkPathCollectorMetricPrefix + "schedule.conns_skipped"
-	standardAllowancePerHour            = 5
+	netpathConnsSkippedMetricName = common.NetworkPathCollectorMetricPrefix + "schedule.conns_skipped"
 )
 
 var getVPCSubnetsForHost = network.GetVPCSubnetsForHost
@@ -495,8 +494,7 @@ func (s *npCollectorImpl) runTracerouteForPath(ptest *pathteststore.PathtestCont
 	path.TestConfigName = ptest.Pathtest.TestConfigName
 	path.TestConfigSource = ptest.Pathtest.TestConfigSource
 	path.DynamicTestProfile = ptest.Pathtest.DynamicTestProfile
-	path.InAllowance = path.DynamicTestProfile == payload.DynamicTestProfileBasic ||
-		(path.DynamicTestProfile == payload.DynamicTestProfileStandard && s.takeStandardAllowance(s.TimeNowFn()))
+	path.InAllowance = s.inAllowance(path.DynamicTestProfile)
 	path.Tags = ptest.Pathtest.Tags
 	path.SourceProduct = s.collectorConfigs.sourceProduct
 	if path.Origin == payload.PathOriginNetflow {
@@ -707,20 +705,4 @@ func (s *npCollectorImpl) runWorker(workerID int) {
 			}
 		}
 	}
-}
-
-// takeStandardAllowance returns true for the first standardAllowancePerHour
-// completed standard runs in each hour. The hour starts on the first take.
-func (s *npCollectorImpl) takeStandardAllowance(now time.Time) bool {
-	s.allowanceMu.Lock()
-	defer s.allowanceMu.Unlock()
-	if s.allowanceUntil.IsZero() || !now.Before(s.allowanceUntil) {
-		s.allowanceUntil = now.Add(time.Hour)
-		s.allowanceLeft = standardAllowancePerHour
-	}
-	if s.allowanceLeft == 0 {
-		return false
-	}
-	s.allowanceLeft--
-	return true
 }
